feature: PlaceDetail과 PlaceSearch 페이지가 아닌 모달로 분리
